MAGETWO-7740 convertToBytes() fails to parse tasklist output GitHub #243 / #237
On Windows systems in other languages than English, the method
'Magento_Test_Helper_Memory::getWinProcessMemoryUsage()' calls
'Magento_Test_Helper_Memory::convertToBytes()' with an invalid
argument, because the column names from the output of 'tasklist'
are language specific.

The alternative usage of the column index seems to be stable over
the various versions of Windows, where 'tasklist' is available.