Implemented MAGETWO-13986: Create grouping elements
We need to add the following grouping elements:
<parameters>
<containers>
<options>

This is needed to allow proper merging of XML files.  Without these
changes there will be issues when nodes are appended during a merge, where
the appends are performed in such a way as to violate the sequence defined
in the XSD file.  This commit alters the XSD to modify most <xs:sequence>
definitions to be a <xs:all> definition instead.  That will avoid the
append issue during merge.

Since <xs:any> only allows 0 or 1 occurance of an element, we now add a
grouping element to the <xs:any> which itself has a <xs:sequence> of only
one element type, which can occur 0 to unbounded number of times.

In the cases where we had an abstract element which was extended, such as
our parameterType, we need to continue using <xs:sequence> even though we
would like to alter it to become <xs:any>, due to the way in which XSD
extension works.  As long as the <xs:sequence doesn't allow for a
maxOccur greater than 1, there shouldn't be a problem with this.