MAGETWO-21567: saveAction of Adminhtml Index controller
Complete refacoring to use services.  As a result the services themselves
had to be refactored.  This work included simplifying the services
to take less parameters, but to handle more logic related to deleting
addresses for example.

Removed any backend logic related to setting password since the FE removed
this ability in MAGETWO-9619.