MAGETWO-22414: Fix resource upgrade version number checking logic
There are a couple of issues here.  One is that we have upgrade
scripts with typo's in them that makes them try to upgrade from a
higher version number to a lower version number.  Another issue is
that the code was attempting to run these scripts when they should
be skipped as they really represent a downgrade of version numbers.