MAGETWO-22741: Product\Composite render function must read CurrentCustomerId
Changes had been made based on the assumption that the Cart controller was correct
since it was the only failing test.  It turns out the past bug was actually in the
cart controller and changes made to Product\Composite broke other controllers.

Changes here should fix existing controllers and adds a test to help verify we
don't hit this again.