MAGETWO-24617: Modify updateTaxRate interface to return bool
Making this change to be more inline with our existing customer services.
There will be extra network overhead to return a data object; if the
update request succeeds then there should be no need to return a data
object, as it should match what the input data object for the update
method was.