MAGETWO-26539: Removed BrowserNotSupportedException
This exception is currently never thrown and it may end up never being
thrown.  There is no obvious security benefit from throwing it.  There
will never be a guarantee that it is always thrown when approriate as
detecting browser capabilities will require JS or a dictionary of all
browsers lacking features (current and future)